+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/Constants.kt
@@ -37,7 +37,7 @@ object Constants {
         const val CLOSEST_VERSION = "closest"
         const val LATEST_VERSION = "latest"
         const val PLATFORM_VERSION = "platform"
-        val MINIMAL_GRADLE_VERSION: GradleVersion = GradleVersion.version("8.5")
+        val MINIMAL_GRADLE_VERSION: GradleVersion = GradleVersion.version("8.6")
         val MINIMAL_INTELLIJ_PLATFORM_BUILD_NUMBER = "223".toVersion()
         val MINIMAL_INTELLIJ_PLATFORM_VERSION = "2022.3".toVersion()
         val MINIMAL_SPLIT_MODE_BUILD_NUMBER = "241.14473".toVersion()