+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/providers/JavaRuntimeMetadataValueSource.kt
@@ -9,9 +9,9 @@ import org.gradle.process.ExecOperations
 import org.jetbrains.intellij.platform.gradle.models.ProductInfo
 import org.jetbrains.intellij.platform.gradle.models.launchFor
 import org.jetbrains.intellij.platform.gradle.utils.asPath
+import org.jetbrains.intellij.platform.gradle.utils.safePathString
 import java.io.ByteArrayOutputStream
 import javax.inject.Inject
-import kotlin.io.path.pathString
 
 /**
  * Obtains the architecture of the provided Java Runtime executable by requesting the list of its internal properties.
@@ -98,7 +98,7 @@ abstract class JavaRuntimeMetadataValueSource : ValueSource<Map<String, String>,
     override fun obtain() = ByteArrayOutputStream().use { os ->
         execOperations.exec {
             commandLine(
-                parameters.executable.asPath.pathString,
+                parameters.executable.asPath.safePathString,
                 "-XshowSettings:properties",
                 "-version",
             )