+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/services/IdesManagerService.kt
@@ -8,9 +8,9 @@ import com.jetbrains.plugin.structure.ide.layout.MissingLayoutFileMode.SKIP_SILE
 import org.gradle.api.services.BuildService
 import org.gradle.api.services.BuildServiceParameters
 import org.jetbrains.intellij.platform.gradle.utils.Logger
+import org.jetbrains.intellij.platform.gradle.utils.safePathString
 import java.nio.file.Path
 import java.util.concurrent.ConcurrentHashMap
-import kotlin.io.path.pathString
 
 /**
  * Abstract service class for managing IntelliJ Platform IDE instances configured within Gradle builds.
@@ -21,7 +21,7 @@ abstract class IdesManagerService : BuildService<BuildServiceParameters.None> {
 
     private val ides = ConcurrentHashMap<String, Ide>()
 
-    fun resolve(platformPath: Path) = ides.computeIfAbsent(platformPath.pathString) {
+    fun resolve(platformPath: Path) = ides.computeIfAbsent(platformPath.safePathString) {
         log.info("Creating new IDE instance from: $platformPath")
         createIde {
             missingLayoutFileMode = SKIP_SILENTLY