+++ src/test/kotlin/org/jetbrains/intellij/platform/gradle/tasks/PrintProductsReleasesTaskTest.kt
@@ -26,6 +26,7 @@ class PrintProductsReleasesTaskTest : IntelliJPluginTestBase() {
             assertContains(
                 """
                 > Task :${Tasks.PRINT_PRODUCTS_RELEASES}
+                IC-2025.2
                 IC-2025.1.4
                 IC-2024.3.6
                 IC-2024.2.6